report: Take a list of input files as command line arguments
Instead of --datadir, there is --staticdatadir for overriding
the directory with colors and test descriptions.

Result files are taken as command line arguments.

+++ report/build.py
@@ -1,5 +1,8 @@
 #! /usr/bin/env python3
 
+"""Generate a HTML report based on network test output
+"""
+
 import os
 import json
 import shutil
@@ -11,13 +14,12 @@ import click
 import yaml
 
 
-SOURCE_FILES = {
-    'results': 'results/*.json',
+STATIC_DATA_FILES = {
     'schema': 'test-client-server-schema.yaml',
     'result_colors': 'result-colors.yaml',
 }
 
-def load_file(filename):
+def load_data(filename):
     print('Loading', filename, file=sys.stderr)
     if filename.endswith('.json'):
         with open(filename, encoding='utf-8') as file:
@@ -28,43 +30,45 @@ def load_file(filename):
     else:
         raise ValueError('Unknown extension: ' + filename)
 
-def load_data(filename_base):
-    if '*' in filename_base:
-        result = {}
-        filenames = glob.glob(filename_base)
-        if not filenames:
-            raise FileNotFoundError(filename_base)
-        for filename in filenames:
-            data = load_file(filename)
-            try:
-                data = dict(data)
-            except ValueError:
-                raise ValueError('Data needs to be a mapping: ', filename)
-            if not data.keys().isdisjoint(result.keys()):
-                raise ValueError('Overwriting data: ', filename)
-            result.update(data)
-        return result
-    else:
-        return load_file(filename_base)
-
 base_path = os.path.dirname(__file__)
 
-@click.command()
-@click.option('-d', '--datadir', default=os.path.join(base_path, 'example_data/'),
-              help='Directory containing input data (defaults to included example data)')
+@click.command(help=__doc__)
+@click.option('-s', '--staticdatadir', default=os.path.join(base_path, 'static_data/'),
+              help='Alternate directory containing static data such as test descriptions')
 @click.option('-o', '--outdir', default=os.path.join(base_path, 'output/'),
               help='Directory for output (will be overwritten)')
 @click.option('--templatedir', default=os.path.join(base_path, 'templates/'),
               help='Directory containing HTML templates')
-def build(datadir, outdir, templatedir):
+@click.argument('input_file', nargs=-1)
+def build(staticdatadir, outdir, templatedir, input_file):
+    if not input_file:
+        raise click.UsageError('No files given. Try running: {} {}'.format(
+              sys.argv[0], os.path.join(base_path, 'example_data/*')))
+        print(__doc__, file=sys.stderr)
+        print(file=sys.stderr)
+        print('No files given. Try running:',
+              sys.argv[0], os.path.join(base_path, 'example_data/*'),
+              file=sys.stderr)
+        return
+
     try:
         shutil.rmtree(outdir)
     except FileNotFoundError:
         pass
     os.makedirs(outdir)
 
-    data = {ident: load_data(os.path.join(datadir, filename))
-            for ident, filename in SOURCE_FILES.items()}
+    data = {ident: load_data(os.path.join(staticdatadir, filename))
+            for ident, filename in STATIC_DATA_FILES.items()}
+    data['results'] = results = {}
+    for filename in input_file:
+        result = load_data(filename)
+        try:
+            result = dict(result)
+        except ValueError:
+            raise ValueError('Data needs to be a mapping: ', filename)
+        if not result.keys().isdisjoint(results.keys()):
+            raise ValueError('Overwriting data: ', filename)
+        results.update(result)
 
     loader = jinja2.FileSystemLoader(templatedir)
     template_env = jinja2.Environment(
